branches/zip: Do not update the free bits in the insert buffer bitmap when inserting or deleting from the insert buffer B-tree.  Assert that records in the insert buffer B-tree are never updated.  This could cure Issue #135.
btr_cur_optimistic_insert(): Do not update the insert buffer bitmap
when inserting to the insert buffer tree.

btr_cur_optimistic_delete(): Do not update the insert buffer bitmap
when deleting from the insert buffer tree.  This could be the cause
of the assertion failure that was reported in Issue #135.

btr_cur_update_alloc_zip(): Assert that the index is not the insert
buffer.  The insert buffer will never be stored in compressed format.

btr_cur_update_in_place(), btr_cur_optimistic_update(),
btr_cur_pessimistic_update(): Assert that these functions are never
invoked on the insert buffer tree.  The insert buffer only supports
the insertion and deletion of records.