Bug#42027: Incorrect parsing of debug and verbose options for mysqldumpslow
Options got normalised to long rather than short options
since we gave primary name and alias in wrong order.
Consequently querying for the option using the short
options (the correct primary name) didn't work, rendering
the options in question inaccessible.

We restore the right order of the universe, or at least
the alii for --debug and --verbose.

scripts/mysqldumpslow.sh:
  Normalise --verbose/-v and --debug/-d to short
  options, not long options.