Bug#42857 Got error 4541 Indexbound has no bound information from NDBCLUSTER
MySQLD calls NdbIndexScanOperation::setBound() with an empty ('open range')
IndexBound with the query given.

setBound() is modified to handle this by inserting an 'open range' bound
which is actually a lower bound of LE NULL on the first column in the index.

The example testcase is added to ndb_index_ordered.