Bug#42238 NDBAPI : NdbRecord insert permits key column value ambiguity
The NdbRecord Insert Api allows key columns to be specified in the key record
and in the attribute record.  It is important that only one key column value 
for each key column is sent to the kernel.
This patch ensures this by taking key column values from the key record for
insert and write operations.
For scan takeover update operations, key column values are taken from the
attribute record.
The testNdbApi testcase is extended with 3 new testcases : 
 - NdbRecordPKAmbiguity
   Verifies that rows inserted/written with divergent key values in the
   key and attribute records contain the correct key values
 - NdbRecordPKUpdate
   Verifies that PK columns can be updated to the same values they contain
   via PK and scan takeover update operations.
 - NdbRecordCICharPKUpdate
   Verifies that a PK column with a case-insensitive VARCHAR column can
   be updated to contain distinct but collation-equal values using PK
   and scan takeover update operations.

These testcases are added to daily-basic-tests.