Increase testBlobs robustness
testBlobs often fails when running against blobs-on-disk tables with
timeout errors.
This occurs as the deadlock detection timeout is fixed per-TC block,
but reading/writing data to a heavily loaded disk subsystem can take
a long time and trigger the timeout.
Potential fixes : 
 1 Increase deadlock detection timeout for all hosts where ATRT is run
 2 Make deadlock detection timeout configurable from API on per transaction
   or operation basis
 3 Improve robustness of testBlobs TC, allowing bounded retries in case of 
   timeouts.

This patch takes approach 3.  Approach 2 should be considered in future.

Additionally, the number of rows used for the testcase is reduced from 100
to 25, reflecting the 4-fold increase in testcases introduced by testing
NdbRecAttr + NdbRecord APIs + Memory + Disk based Blobs.  This should reduce
the runtime of the testBlobs testcases.